refactor: move layerCreationFormDialog to features/layers
