+++ packages/base/src/commands/index.ts
@@ -20,8 +20,8 @@ import { fromLonLat } from 'ol/proj';
 
 import { addLayerCreationCommands } from './operationCommands';
 import { CommandIDs, icons } from '../constants';
-import { LayerCreationFormDialog } from '../dialogs/layerCreationFormDialog';
 import { LayerBrowserWidget } from '../features/layer-browser';
+import { LayerCreationFormDialog } from '../features/layers/layerCreationFormDialog';
 import { SymbologyWidget } from '../features/layers/symbology/symbologyDialog';
 import { ProcessingFormDialog } from '../features/processing/ProcessingFormDialog';
 import { getSingleSelectedLayer } from '../features/processing/index';
