+++ packages/base/src/index.ts
@@ -1,7 +1,7 @@
 export * from './features/annotations';
 export * from './commands/index';
 export * from './constants';
-export * from './dialogs/layerCreationFormDialog';
+export * from './features/layers/layerCreationFormDialog';
 export * from './icons';
 export * from './mainview';
 export * from './menus';
